Добавлена отрисовка оружия
