+++ Data/Levels/lvl_1.lvl
@@ -11,7 +11,7 @@
 -                    ----
 ------               ----
 ------               ----
-------               ----
+------        G      ----
 ------       ----    ----
 ------       ----    ----
 ------       ----    ----
