+++ Data/Sprites/Player/goes_right2.png
