+++ Data/Sprites/Player/jump_right.png
