+++ Data/Sprites/Player/stand_right.png
