+++ Data/Sprites/Player/stand_sraight.png
